pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autopwn-ctf"
version = "3.1.0"
description = "Autonomous CTF pwn challenge solver with multi-phase analysis pipeline"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "AutoPwn Contributors" },
]
keywords = ["ctf", "pwn", "exploit", "binary", "security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "pwntools",
    "ROPgadget",
    "LibcSearcher",
]

[project.optional-dependencies]
full = [
    "angr",
]

[project.scripts]
autopwn = "autopwn.cli:main"

[project.urls]
Homepage = "https://github.com/Lx-autopwn/AutoPwn"
Repository = "https://github.com/Lx-autopwn/AutoPwn"

[tool.setuptools.packages.find]
include = ["autopwn*"]
